pyproject.toml
[project]
name = "ms-boilerplate"
version = "0.1.0"
description = "A simple ms boilerplate for Python"
authors = [
    {name = "Hari S3ldon",email = "enrico.pascucci@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard] (>=0.119.0)",
    "pydantic (>=2.12.0)",
    "sqlalchemy (>=2.0.0)",
    #"opentelemetry-instrumentation-fastapi",
    #"opentelemetry-instrumentation-sqlalchemy",
    "asyncpg (>=0.30.0)",
    "alembic (>=1.17.0)",
    "python-lsp-server[all] (>=1.13.1,<2.0.0)",
    "pylsp-mypy (>=0.7.0,<0.8.0)",
    "python-lsp-black (>=2.0.0,<3.0.0)",
    "python-lsp-isort (>=0.2.1,<0.3.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest (>=7.0.0)",
    "httpx (>=0.23.0)",
    "pytest-asyncio (>=0.18.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
